requirements.txt
binance==1.0.10
python-dotenv==1.0.0
matplotlib==3.7.1
openpyxl==3.1.2
requests==2.31.0
nest-asyncio==1.5.6
pytz==2023.3
plyer==2.1.0
